POC running features in parallel using t.Parallel
This is a proof of concept to show that we can run features in parallel
using the `t.Parallel()` function from the `testing` package instead of
relying on the custom `--parallel` flag provided by the framework.
The issue with the current implementation is that using `t.Parallel()`
in 2 features of the same test yields the error `panic: testing:
t.Parallel called multiple times`.
This POC tries to solve it by running each feature in its separate
subtest instance.