[config] Refactor how config's client is generated
Currently, the config's client is managed as follow:
- it can be created with `NewClient` and `Client`. However, once it is
created, it is cached and will always be returned.
- it can be set directly with `WithClient`.

The current approach has a few drawbacks when running tests in parallel:
- the client is shared between tests, leading to race conditions when
it is initialized at the same time in multiple tests.
- because of the caching, if one gets the client, modifies the kubeconfig
or any other client related parameter in the config, and then gets the
client again, they would get the original client and not a new one with
the new kubeconfig.
- `Client` is not a simple getter function like all the other similar
methods of the config, which makes it confusing to users.

This commit attempts to fix this by initializing the client earlier but
currently does not work for in-cluster clients.
It conflicts with changes made in #57.